Fix: functions signtaures and some typing

+++ genetic_algo/fitness.py
@@ -6,7 +6,7 @@ from numba import prange
 from genetic_algo.lgfi import lgfi
 from genetic_algo.solutions_generation import get_corresponding_sequence_by_id
 
-@njit(float32(from_dtype(Bin)), cache = True)
+@njit(float64(from_dtype(Bin)), cache = True)
 def calculate_bin_fill(bin: np.ndarray) -> int:
     """
     Calculate the total fill of a bin based on the items placed in it.
@@ -28,7 +28,7 @@ def calculate_bin_fill(bin: np.ndarray) -> int:
         
     return total_fill / (bin['width']*bin['height'])
 
-@njit(float32(from_dtype(Item)[:], int32[:], UniTuple(int32, 2), boolean, boolean), cache = True)
+@njit(float64(from_dtype(Item)[:], int32[:], UniTuple(int32, 2), boolean, boolean), cache = True)
 def compute_fitness(items: np.ndarray, id_ordering: np.ndarray, bin_dimensions: Tuple[int, int], 
                     guillotine_cut: bool, rotation: bool):
     """
@@ -58,7 +58,7 @@ def compute_fitness(items: np.ndarray, id_ordering: np.ndarray, bin_dimensions: 
     solution_fitness = len(solution) + calculate_bin_fill(solution[-1])
     return solution_fitness
 
-@njit(float32[:](int32[:, :], from_dtype(Item)[:], UniTuple(int32, 2), boolean, boolean), parallel = True, cache = True)
+@njit(float64[:](int32[:, :], from_dtype(Item)[:], UniTuple(int32, 2), boolean, boolean), parallel = True, cache = True)
 def compute_fitnesses(population: np.ndarray, items: np.ndarray, bin_dimensions: Tuple[int, int], 
                       guillotine_cut: bool, rotation: bool) -> np.ndarray:
     """
@@ -77,7 +77,7 @@ def compute_fitnesses(population: np.ndarray, items: np.ndarray, bin_dimensions:
     
     population_size = len(population)
     
-    fitnesses = np.zeros(population_size, dtype=float32)
+    fitnesses = np.zeros(population_size, dtype=np.float64)
     
     for i in prange(population_size):
